Remove package kineprod, since it’s not extractable as is.
